Updating server error resolution

+++ server/DataCollection/src/main/java/com/cdac/Acts/Repositories/SentenceRepository.java
@@ -8,7 +8,9 @@ import org.springframework.stereotype.Repository;
 import com.cdac.Acts.entities.Sentence;
 
 @Repository
-public interface SentenceRepository extends JpaRepository<Sentence, Long> {
-    List<Sentence> findByDocumentId(Long documentId);
+public interface SentenceRepository extends JpaRepository<Sentence, Integer> {
+
+	List<Sentence> findByDocumentId(int documentId);
+   // List<Sentence> findByDocumentId(Long documentId);
     
 }
