+++ server/DataCollection/src/main/java/com/cdac/Acts/Repositories/UserRepository.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.cdac.Acts.entities.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUsername(String username);
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+	User findByUsername(String username);
+   // User findByUsername(String username);
 }
